% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnvFunMarkCor.R
\name{EnvFunMarkCor}
\alias{EnvFunMarkCor}
\title{Function to obtain the Functional Mark Correlation Function}
\usage{
EnvFunMarkCor(
  xy,
  r,
  rf,
  funpp,
  stoyan = 0.15,
  kernel = "epanechnikov",
  bw,
  lambda,
  correction = "isotropic",
  simul = 299,
  alpha = 0.05,
  verbose = TRUE,
  Figure = TRUE,
  miny,
  maxy
)
}
\arguments{
\item{xy, }{object of class ppp (point pattern)}

\item{r, }{sequence of values to evaluate the Functional mark correlation function}

\item{rf, }{sequence of values for the function associated to each point}

\item{funpp}{is a matrix that contains de functions for each point, funct\link{,1} is the function for point 1}

\item{stoyan}{value for the epanechnikov kernel, default stoyan=0.15}

\item{kernel, }{kernel function for both the pair correlation function and the Functional mark correlation function}

\item{bw}{bandwidth for the kernel function for the pair correlation function and Functional mark correlation function}

\item{lambda, }{point intensity}

\item{correction}{for edge-effect for both the pair correlation function and the Functional mark correlation function}

\item{simul, }{number of permutations for the Monte Carlo test, default simul=999}

\item{alpha, }{signification level of the Monte Carlo Test}

\item{verbose, }{show the simulation progression, default verbose=TRUE}

\item{Figure, }{display resulting FMCF and related max and min evelops. default Figure=TRUE}

\item{miny, }{maxy, plot parameters, ylim=c(miny,maxy)}
}
\value{

}
\description{
This function computes the Functional Mark Correlation Function as described in Comas et al (2009) and Comas et al. (2013)
We consider as a test function for the functional point pattern, L2 distances between functions
}
\examples{
  To be implemented
 
}
\author{
Carles Comas \email{carles.comas@udl.cat}
}
